Allow to use a proxy for downloads
This change will make it possible to provide a proxy when downloading.

For that the content of the environment-variable `HTTP_PROXY` is used.
Alternatively one can also use `http_proxy` (as used by curl) can be
used.

For proxy-authentication you can set user and password for basic-
authentication via the environment variable `HTTP_PROXY_AUTH`